% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SClineager_eda.R
\name{explore_sclineager}
\alias{explore_sclineager}
\title{Exploratory analysis by visualization}
\usage{
explore_sclineager(folders, coverage_cutoff, file)
}
\arguments{
\item{folders}{A path where each subfolder has the raw output from one \link{read_sclineager} call (for cells of one sample), which contains the ``cleaned.RData'' file.}

\item{coverage_cutoff}{Coverage below this cutoff is considered not observable.}

\item{file}{A file name (extension pdf) at which outputs (figures) are saved. For example, ``figure_eda.pdf''.}
}
\value{
This returns nothing, but figures in \code{file}.
}
\description{
Explore data before applying SClineager.
}
