% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SClineager_fitting.R
\name{run_sclineager}
\alias{run_sclineager}
\title{An wrapper function for running SClineager}
\usage{
run_sclineager(file_in, folder, categories, max_iter, mask_genes,
  vaf_offset, dfreedom, skip_common, psi = NULL, control = NULL,
  save = FALSE)
}
\arguments{
\item{file_in}{A path pointing the input file of the .RData file generated by \link{read_sclineager}.
. It must have a list named "results" with components "runinfo", "mutations_mat", "annotation", and "coverage_mat".}

\item{folder}{A folder at which outputs are saved. If it does not exist, it will be automatically created.}

\item{categories}{A character vector consisting of any combination of ``frameshift substitution'', ``nonframeshift substitution'', ``nonsynonymous SNV'', ``synonymous SNV'', ``splicing'', ``UTR5'', ``UTR3'', ``stopgain'', and ``stoploss''.}

\item{max_iter}{The number of posterior samples. Half of samples are discarded as burn-in steps.}

\item{mask_genes}{A character vector. Genes in this vector will be masked from the analyses. It is probably a good idea to mask all HLA genes from the analysis.}

\item{vaf_offset}{A small number (e.g. 1e-2) for converting VAF to logit VAF.}

\item{dfreedom}{(Hyperparameter) the degrees of freedom used in inverse Wishart distribution of \code{psi}. Should be larger than \code{ncol(mutations_mat)-1}.}

\item{skip_common}{For the variants in consideration, if this flag is set to TRUE, the variants that are common in the human population (esp6500siv2_all>0.01) will be skipped.}

\item{psi}{(Hyperparameter) covariance matrix of VAF across samples. If \code{NULL} (default), it will be set to an identity matrix.}

\item{control}{Control parameters for determining the variance function relating the observed VAF to the true VAF (see Details). If \code{NULL} (default), then a=1, b=0.5, c=1, d=20, e=0.4.}

\item{save}{Logical. If TRUE, posterior samples are saved and returned. Default is \code{FALSE}.}
}
\value{
``imputation_results.pdf'' and ``results.RData'' are saved in \code{folder}. The .RData file contains a list with components (also returned):
\item{genotype_mat}{(Parameter) an estimated VAF matrix, obtained by averaging last half of posterior samples.}
\item{genotype_mat_all}{A list of posterior samples for VAF. \code{NULL} is returned if \code{save} is \code{FALSE}.}
\item{genotype_mat_orig}{Same as \code{mutations_mat}, but contains the raw VAF data.}
\item{sigma}{(Parameter) an estimated covariance matrix of VAF across samples, obtained by averaging last half of posterior samples.}
\item{sigma_all}{A list of posterior samples for a covariance matrix.  \code{NULL} is returned if \code{save} is \code{FALSE}.}
\item{k_mat}{A matrix whose entry is a variance function computed by empirical estimates.}
\item{loglike}{A vector of posterior likelihoods.}
\item{runinfo}{Same as \code{runinfo} in the input, but trimmed to contain only cells that appeared in \code{mutations_mat}, in the same order.}
\item{annotation}{Annotation information of the variants in \code{mutations_mat}.}
}
\description{
This function wraps the main function \link{sclineager_internal} to process data and return outputs.
}
\examples{
# See the github page https://github.com/inmybrain/SClineager.
}
