% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SClineager_fitting.R
\name{sclineager_internal}
\alias{sclineager_internal}
\title{The Gibbs sampler for SClineager}
\usage{
sclineager_internal(
  mutations_mat,
  coverage_mat,
  max_iter,
  vaf_offset,
  dfreedom,
  psi,
  control = NULL,
  save = FALSE
)
}
\arguments{
\item{mutations_mat}{(Data) The variant allele frequency (VAF) matrix, whose rows are mutations and columns are samples. Each value is in [0,1] and NA value is for unobservable cells.}

\item{coverage_mat}{(Data)  The total sequencing coverage matrix, whose rows are mutations and columns are samples. Each value is positive real number and NA value is for unobservable cells.}

\item{max_iter}{See \link{run_sclineager}.}

\item{vaf_offset}{See \link{run_sclineager}.}

\item{dfreedom}{See \link{run_sclineager}.}

\item{psi}{See \link{run_sclineager}.}

\item{control}{See \link{run_sclineager}.}

\item{save}{See \link{run_sclineager}.}
}
\value{
A list with components:
\item{genotype_mat}{See \link{run_sclineager}.}
\item{genotype_mat_all}{See \link{run_sclineager}.}
\item{genotype_mat_orig}{See \link{run_sclineager}.}
\item{sigma}{See \link{run_sclineager}.}
\item{sigma_all}{See \link{run_sclineager}.}
\item{k_mat}{See \link{run_sclineager}.}
\item{loglike}{See \link{run_sclineager}.}
}
\description{
Perform the MCMC sampling to estimate parameters in SClineager.
}
\examples{
# See the github page https://github.com/inmybrain/SClineager.
}
\seealso{
\link{run_sclineager}
}
